dodao za filter ali nije gotovo

+++ app/src/main/java/rs/raf/rma/nutritiontrackerrma/data/datasources/local/dao/ListMealDao.kt
@@ -18,7 +18,7 @@ abstract class ListMealDao {
     abstract fun getAll(): Observable<List<ListMealEntity>>
 
     @Query("SELECT * FROM meals WHERE calories BETWEEN :minCalories AND :maxCalories ORDER BY calories DESC")
-    abstract fun getAllSortedByCalDes(minCalories: Int, maxCalories: Int): Observable<List<ListMealEntity>>
+    abstract fun getAllSortedByCalDes(minCalories: Double, maxCalories: Double): Observable<List<ListMealEntity>>
 
     @Query("DELETE FROM meals")
     abstract fun deleteAll()
