+++ app/src/main/java/rs/raf/rma/nutritiontrackerrma/presentation/contracts/MealsContract.kt
@@ -27,7 +27,7 @@ interface MealsContract {
         fun getSingleMeal(mealId : String)
         fun getAllSingleMeals(mealList: ArrayList<ListMeal>)
         fun getAllMeals()
-        fun getAllMealsSortedByCal(min:Int,max:Int)
+        fun getAllMealsSortedByCal(min:Double,max:Double)
         fun getMealByName(name: String)
         fun getAllSavedMeals(user:String)
         fun getSavedMealByName(name: String)
